Add First Working Protoype
The code added in this commit is a big update (as it was not committed
in small chunks going along) which adds the core functionality to
autospice, including:
 - Load settings from a .yml file
 - Pre-allocated code for different simulation codes, machines and
   queueing systems
 - Auto-submission of jobs when all validated
 - Auto-logging of submission into a google sheet which acts as a
   central, cloud-based database

This has had soe basic tests run through it and the dryrun_fl works as
anticipated. Testing now needs to advance to a live test.