Add Parameter Scan Support and Tidy Main Function
Support is added for parameter scans to be run when detected in the
input file. Currently this only works for a scan in one parameter, i.e.
a 1-dimensional parameter scan, but more are planned.

The main submit_job function is also tidied up a bit with the scheduler
configuration options moved to a separate function and directory IO
refactored into a code-specific method.
:x