Fix Exectuable Dir Passing and Add Config Copying
The bug with not properly passing the executable directory to the
submission script body is fixed with the inclusion of an extra parameter
in call_params. The call params are individually extracted in the code
implementation of get_submission_script_body() instead of prescribed by
the parent class.

Also added a command to copy the config file (yaml file) to the output
directory and fixed a code typo where the loop over n_jobs was
improperly implemented.