Add Option to Put Header Parameters on Single Line
Can now optionally specify a list of parameters to be added to the
submission script header (the scheduler directives) on a single line.
This was inspired by the fact that nodes and processors-per-node must be
defined together for the PBS scheduler. This is controlled by the kwarg
join_lines on the get_submission_script_header() method of the Scheduler
class, and implemented currently as a default argument for each subclass
of scheduler.