Improve Handling of Scheduler Options from Yaml
The handling of scheduler options specified in the input file is now
improved to allow arbitrary entering of optional parameters that are
specified in the scheduler definition. If an option is given that is not
defined then a warning will show before confirmation to notify the user.

This has also meant submission_params is now standardised to only
containing parameters from the scheduler definition.