Assign Values for Log Names Before Multisubmission
The default names of the logs are assigned before the multisubmission
logic, as this was previously not being defined in a single submission
case and therefore breaking as log names are required by the scheduler.

Have also extended the machine max job time error message to be slightly
more helpful.