Remove 'user' From Slurm Scheduler Definition
The 'user' parameter is removed from slurm by default as the --uid
command requires root priviliges in slurm. Have also therefore added it
to the ignored parameters so there is no error message every time user
is specified. NOTE: This last point should probably be a list generated
on a per-scheduler basis.