Add Backing-up As Command Line Option
The option to backup the data in the simulation directory is now able to
be disabled at command line. This is SimulationCode specific, but SPICE
uses this to copy contents of sim directory into a new directory once
the main run has finished - usually due to simulation time limits.

Default is on and it can be disabled with a -b or --backup_fl in the
command line call.