Fix Issue with Existing Directory Handling
The changes to restart modes introduced a bug whereby upon creation of a
new directory if the requested one is found to already exist, the
existing directory was still being passed to the submission file
creation functions, which resulted in some data being overwritten.

This has been fixed by properly reassigning the call_params value.