Implement Copying Executable to Output Directory
There is now an additional, optional input to the [code] section of the
yaml file called copy_exe, which controls whether or not the executable
(as well as some additional code-specific files/folders) are copied to
each output dir into some subfolder. The subfolder is named in
codes/core.py.